Remove retry in the dataconn.Client.
The retry in dataconn.Client was creating additional work for the
longhorn replica; if the retry is successful, the I/O operation would be
performed twice.  Changed the timout for reads and writes to eight
seconds instead of four seconds with one retry.

I also made the ping timeout one second instead of eight seconds because
the ping is not disk bound and can be used to quickly determine if the
replica is unresponsive.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>