Stop monitoring goroutine when removing replica
The monitoring goroutine was continuing to run after the replica was
removed.  This caused a problem when a replica is recreated for the same
IP address and port after one is deleted.  When a replica is removed,
the channel that the monitoring goroutine is waiting for will eventually
get an error message.  In the normal case, the SetReplicaMode won't find
a replica with the address to mark as ERR.  However, when another
replica is with the same address is added after the original was
removed, SetReplicaMode for the old monitorChan will mark the new
replica as ERR with SetReplicaMode.

This change makes removing the replica also stop the monitoring
goroutine that was launched for that replica.  It puts a nil error on
the replica's monitor channel so monitoring goroutine  will exit as part
of removing the replica.

Signed-off-by: Keith Lucas <keith.lucas@suse.com>